moved csp filter to views module, seeing couple test failures with the refactor
