+++ http-server-netty/src/test/groovy/io/micronaut/http/server/netty/configuration/NettyHttpServerConfigurationSpec.groovy
@@ -138,32 +138,4 @@ class NettyHttpServerConfigurationSpec extends Specification {
         cleanup:
         beanContext.close()
     }
-
-    void "test csp configuration"() {
-        given:
-        ApplicationContext applicationContext = new DefaultApplicationContext("test")
-        applicationContext.environment.addPropertySource(PropertySource.of("test",
-                ['micronaut.server.csp.enabled': true,
-                 'micronaut.server.csp.reportOnly': false,
-                 'micronaut.server.csp.policyDirectives': "default-src https:"]
-
-        ))
-        applicationContext.start()
-
-        expect:
-        applicationContext.containsBean(HttpServerConfiguration.CspConfiguration)
-
-        when:
-        NettyHttpServerConfiguration config = applicationContext.getBean(NettyHttpServerConfiguration)
-        NettyHttpServer server = applicationContext.getBean(NettyHttpServer)
-        server.start()
-
-        then:
-        config.csp.enabled
-        !config.csp.reportOnly
-        config.csp.policyDirectives == 'default-src https:'
-
-        cleanup:
-        applicationContext.close()
-    }
 }
